lag endepunkt for å hente PDF fra IM-id (#119)

+++ src/main/java/no/nav/familie/inntektsmelding/imdialog/modell/InntektsmeldingRepository.java
@@ -47,7 +47,7 @@ public class InntektsmeldingRepository {
         return query.getResultList();
     }
 
-    public InntektsmeldingEntitet hentInntektsmelding(int inntektsmeldingId) {
+    public InntektsmeldingEntitet hentInntektsmelding(long inntektsmeldingId) {
         return entityManager.find(InntektsmeldingEntitet.class, inntektsmeldingId);
     }
 }
